chore(refactor): extract static + vault routes; finish the index.ts series
Eighth and final route bucket. Three new modules:

server/src/vault-inventory.ts — pure data: walkDirSize / countTopEntries /
humanizeHome / VaultInventoryEntry / buildVaultInventory + the cache.
Public API is getVaultInventory(deps) — first call within
VAULT_INVENTORY_TTL_MS (30s) does the walk; subsequent calls reuse.

server/src/routes/static.ts — seven endpoints, plus resolveArtifactsUrl
(only callers are in this module now):
  GET /api/resolve-path
  GET /api/workspace
  GET /api/vault/inventory
  GET /api/apps/:name/start
  GET /api/apps/:name/stop
  GET /docs/:name
  GET /artifacts/<rel>

server/src/mime.ts — tiny shared MIME map. Both the static route module
and the SPA fallback in index.ts need it; kept in one place so adding a
content type doesn't require touching two files.

Drops from index.ts: walkDirSize, humanizeHome, countTopEntries,
VaultInventoryEntry, vaultInventoryCache, VAULT_INVENTORY_TTL_MS,
invalidateVaultInventoryCache, buildVaultInventory, resolveArtifactsUrl,
the inline MIME table, all 7 inline route handlers, and the now-unused
imports (renderMarkdown, renderMermaid, inferName, injectBridge).

Cumulative across the audit: index.ts is 643 lines — down 1575 lines
(-71%) from the pre-audit baseline of 2218.

No behavioural change.
